intermediate/challenge-1/bus_scheduler/requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.6.1
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.2
black==22.6.0
blinker==1.5
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
codecov==2.1.12
colorama==0.4.5
commonmark==0.9.1
coverage==6.4.2
Deprecated==1.2.13
dnspython==2.2.1
dynaconf==3.1.9
email-validator==1.2.1
fakeredis==1.8.1
fastapi==0.79.0
fastapi-mail==1.0.9
flake8==4.0.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.971
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pyflakes==2.4.0
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
python-multipart==0.0.5
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
rich==12.5.1
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.39
starlette==0.19.1
tomli==2.0.1
typer==0.6.1
typing_extensions==4.3.0
urllib3==1.26.10
uvicorn==0.18.2
wrapt==1.14.1
python-jose==3.3.0